5.4.19-20250822_17558320

+++ polardbx-test/src/test/java/com/alibaba/polardbx/qatest/ddl/sharding/ddl/ForeignKeyDdlTest.java
@@ -1,7 +1,7 @@
 package com.alibaba.polardbx.qatest.ddl.sharding.ddl;
 
+import com.alibaba.polardbx.qatest.CdcIgnore;
 import com.alibaba.polardbx.qatest.DDLBaseNewDBTestCase;
-import com.alibaba.polardbx.qatest.ReplicaIgnore;
 import com.alibaba.polardbx.qatest.util.JdbcUtil;
 import org.junit.After;
 import org.junit.Assert;
@@ -16,7 +16,7 @@ import java.sql.Statement;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-@ReplicaIgnore(
+@CdcIgnore(
     ignoreReason = "目前创建外键的时候会同步创建一个同名的索引，导致了Duplicate Key，导致下游复制中断，暂时忽略")
 public class ForeignKeyDdlTest extends DDLBaseNewDBTestCase {
     private static final String FOREIGN_KEY_CHECKS = "FOREIGN_KEY_CHECKS=TRUE";
@@ -55,6 +55,18 @@ public class ForeignKeyDdlTest extends DDLBaseNewDBTestCase {
         + "    b int not null,\n"
         + "    c int not null,\n"
         + "    d int not null,\n"
+        + "    foreign key (`b`) REFERENCES `user1` (`b`)\n"
+        + ") %s";
+
+    private static final String CREATE_DEVICE_WITH_KEY = "create table device_with_key\n"
+        + "(   a int auto_increment primary key,\n"
+        + "    b int not null,\n"
+        + "    c int not null,\n"
+        + "    d int not null,\n"
+        + "    `TITLE` text NOT NULL COMMENT '标题',"
+        + "    index (b),\n"
+        + "    index (c,d),\n"
+        + "    FULLTEXT KEY (TITLE) WITH PARSER ngram,\n"
         + "    foreign key (`b`) REFERENCES `user1` (`b`)\n"
         + ") %s";
 
@@ -302,6 +314,10 @@ public class ForeignKeyDdlTest extends DDLBaseNewDBTestCase {
         JdbcUtil.executeUpdateSuccess(tddlConnection, "SET ENABLE_FOREIGN_KEY = true");
 
         for (String partitionDef1 : PART_DEFS) {
+            if (partitionDef1.contains("single") || partitionDef1.contains("broadcast")) {
+                continue;
+            }
+
             dropTableIfExists("device");
             dropTableIfExists("user1");
 
@@ -315,6 +331,12 @@ public class ForeignKeyDdlTest extends DDLBaseNewDBTestCase {
 
                 JdbcUtil.executeUpdateFailed(tddlConnection, "DROP TABLE user1", "");
                 JdbcUtil.executeUpdateFailed(tddlConnection, "TRUNCATE TABLE user1", "");
+
+                // truncate table with gsi
+                JdbcUtil.executeUpdateSuccess(tddlConnection,
+                    "alter table user1 add global index `gsi_b` (`b`) dbpartition by hash(`b`)");
+                JdbcUtil.executeUpdateFailed(tddlConnection, "TRUNCATE TABLE user1", "");
+                JdbcUtil.executeUpdateSuccess(tddlConnection, "alter table user1 drop index gsi_b");
             }
         }
     }
@@ -745,6 +767,29 @@ public class ForeignKeyDdlTest extends DDLBaseNewDBTestCase {
             JdbcUtil.executeUpdateSuccess(tddlConnection,
                 String.format(
                     "alter table `device_same` add constraint `device_ibfk_2` foreign key (`b`) REFERENCES `user1` (`c`)",
+                    partitionDef2));
+        }
+    }
+
+    @Test
+    public void testCreateFkWithEmptyNameIndex() {
+        JdbcUtil.executeUpdateSuccess(tddlConnection, "SET ENABLE_FOREIGN_KEY = true");
+
+        dropTableIfExists("device_with_key");
+        dropTableIfExists("user1");
+
+        JdbcUtil.executeUpdateSuccess(tddlConnection, String.format(CREATE_USER1, "dbpartition by hash(a)"));
+
+        for (String partitionDef2 : PART_DEFS) {
+
+            System.out.println(partitionDef2);
+            dropTableIfExists("device_with_key");
+            // create table with fk
+            JdbcUtil.executeUpdateSuccess(tddlConnection, String.format(CREATE_DEVICE_WITH_KEY, partitionDef2));
+            // create table and create fk
+            JdbcUtil.executeUpdateSuccess(tddlConnection,
+                String.format(
+                    "alter table `device_with_key` add constraint `device_ibfk_2` foreign key (`b`) REFERENCES `user1` (`c`)",
                     partitionDef2));
         }
     }
